+++ lib/components/Accordion/Accordion.snippets.tsx
@@ -9,7 +9,7 @@ import { Snippets } from '../private/Snippets';
 
 export const snippets: Snippets = [
   {
-    name: 'Standard',
+    name: 'Large',
     code: source(
       <Accordion>
         <AccordionItem label="Item 1">
@@ -25,9 +25,65 @@ export const snippets: Snippets = [
     ),
   },
   {
-    name: 'Standalone item',
+    name: 'Large, without dividers',
+    code: source(
+      <Accordion dividers={false}>
+        <AccordionItem label="Item 1">
+          <Placeholder height={100} />
+        </AccordionItem>
+        <AccordionItem label="Item 2">
+          <Placeholder height={100} />
+        </AccordionItem>
+        <AccordionItem label="Item 3">
+          <Placeholder height={100} />
+        </AccordionItem>
+      </Accordion>,
+    ),
+  },
+  {
+    name: 'Standard',
+    code: source(
+      <Accordion size="standard">
+        <AccordionItem label="Item 1">
+          <Placeholder height={100} />
+        </AccordionItem>
+        <AccordionItem label="Item 2">
+          <Placeholder height={100} />
+        </AccordionItem>
+        <AccordionItem label="Item 3">
+          <Placeholder height={100} />
+        </AccordionItem>
+      </Accordion>,
+    ),
+  },
+  {
+    name: 'Standard, without dividers',
+    code: source(
+      <Accordion size="standard" dividers={false}>
+        <AccordionItem label="Item 1">
+          <Placeholder height={100} />
+        </AccordionItem>
+        <AccordionItem label="Item 2">
+          <Placeholder height={100} />
+        </AccordionItem>
+        <AccordionItem label="Item 3">
+          <Placeholder height={100} />
+        </AccordionItem>
+      </Accordion>,
+    ),
+  },
+  {
+    name: 'Large standalone item',
     code: source(
       <AccordionItem label="Label">
+        <Placeholder height={100} />
+      </AccordionItem>,
+    ),
+  },
+  {
+    name: 'Standard standalone item',
+    code: source(
+      <AccordionItem label="Label" size="standard">
         <Placeholder height={100} />
       </AccordionItem>,
     ),