+++ lib/components/Accordion/AccordionItem.tsx
@@ -1,9 +1,9 @@
-import React, { ReactNode } from 'react';
+import React, { ReactNode, useContext } from 'react';
 import assert from 'assert';
 import { Box } from '../Box/Box';
+import { Text, TextProps } from '../Text/Text';
 import { Columns } from '../Columns/Columns';
 import { Column } from '../Column/Column';
-import { Heading } from '../Heading/Heading';
 import { IconChevron } from '../icons';
 import {
   useDisclosure,
@@ -14,13 +14,29 @@ import {
 import { useVirtualTouchable } from '../private/touchable/useVirtualTouchable';
 import { hideFocusRingsClassName } from '../private/hideFocusRings/hideFocusRings';
 import { Overlay } from '../private/Overlay/Overlay';
+import {
+  AccordionContext,
+  AccordionContextValue,
+  validTones,
+} from './AccordionContext';
+import buildDataAttributes, {
+  DataAttributeMap,
+} from '../private/buildDataAttributes';
 import * as styles from './AccordionItem.css';
 
-const accordionSpace = 'large';
+const itemSpaceForSize = {
+  xsmall: 'small',
+  small: 'medium',
+  standard: 'medium',
+  large: 'large',
+} as const;
 
 export type AccordionItemBaseProps = {
   label: string;
   children: ReactNode;
+  size?: TextProps['size'];
+  tone?: AccordionContextValue['tone'];
+  data?: DataAttributeMap;
 };
 
 export type AccordionItemProps = AccordionItemBaseProps & UseDisclosureProps;
@@ -30,8 +46,34 @@ export const AccordionItem = ({
   id,
   label,
   children,
+  size: sizeProp,
+  tone: toneProp,
+  data,
   ...restProps
 }: AccordionItemProps) => {
+  const accordionContext = useContext(AccordionContext);
+
+  assert(
+    !(accordionContext && sizeProp),
+    'Size cannot be set on AccordionItem when inside Accordion. Size should be set on Accordion instead.',
+  );
+  assert(
+    !(accordionContext && toneProp),
+    'Tone cannot be set on AccordionItem when inside Accordion. Tone should be set on Accordion instead.',
+  );
+
+  assert(
+    toneProp === undefined || validTones.includes(toneProp),
+    `The 'tone' prop should be one of the following: ${validTones
+      .map((x) => `"${x}"`)
+      .join(', ')}`,
+  );
+
+  const size = accordionContext?.size ?? sizeProp ?? 'large';
+  const tone = accordionContext?.tone ?? toneProp ?? 'neutral';
+  const weight = 'medium';
+  const itemSpace = itemSpaceForSize[size] ?? 'none';
+
   assert(
     typeof label === 'undefined' || typeof label === 'string',
     'Label must be a string',
@@ -50,7 +92,7 @@ export const AccordionItem = ({
   });
 
   return (
-    <Box>
+    <Box {...(data ? buildDataAttributes(data) : undefined)}>
       <Box position="relative" display="flex">
         <Box
           component="button"
@@ -67,19 +109,21 @@ export const AccordionItem = ({
             https://stackoverflow.com/questions/41100273/overflowing-button-text-is-being-clipped-in-safari
           */}
           <Box position="relative">
-            <Columns space={accordionSpace}>
+            <Columns space={itemSpace}>
               <Column>
-                <Heading component="div" level="4">
+                <Text size={size} weight={weight} tone={tone} component="div">
                   {label}
-                </Heading>
+                </Text>
               </Column>
               <Column width="content">
-                <Heading component="div" level="4">
-                  <IconChevron
-                    tone="secondary"
-                    direction={expanded ? 'up' : 'down'}
-                  />
-                </Heading>
+                <Text
+                  size={size}
+                  weight={weight}
+                  tone={tone === 'neutral' ? 'secondary' : tone}
+                  component="div"
+                >
+                  <IconChevron direction={expanded ? 'up' : 'down'} />
+                </Text>
               </Column>
             </Columns>
           </Box>
@@ -92,7 +136,7 @@ export const AccordionItem = ({
         />
       </Box>
       <Box
-        paddingTop={accordionSpace}
+        paddingTop={itemSpace}
         display={expanded ? 'block' : 'none'}
         {...contentProps}
       >