+++ lib/components/Actions/Actions.tsx
@@ -6,14 +6,15 @@ import { PrivateButtonRendererProps } from '../ButtonRenderer/ButtonRenderer';
 export interface ActionsProps {
   size?: PrivateButtonRendererProps['size'];
   children: InlineProps['children'];
+  data?: InlineProps['data'];
 }
 
-export const Actions = ({ size, children }: ActionsProps) => {
+export const Actions = ({ size, data, children }: ActionsProps) => {
   const contextValue = useMemo(() => ({ size }), [size]);
 
   return (
     <ActionsContext.Provider value={contextValue}>
-      <Inline collapseBelow="tablet" space="xsmall">
+      <Inline collapseBelow="tablet" space="xsmall" data={data}>
         {children}
       </Inline>
     </ActionsContext.Provider>