+++ lib/components/Columns/Columns.tsx
@@ -11,6 +11,9 @@ import {
   normalizeResponsiveValue,
   ResponsiveValue,
 } from '../../sprinkles/sprinkles.css';
+import buildDataAttributes, {
+  DataAttributeMap,
+} from '../private/buildDataAttributes';
 
 type CollapsibleAlignmentChildProps = ReturnType<
   typeof resolveCollapsibleAlignmentProps
@@ -40,6 +43,7 @@ export interface ColumnsProps extends CollapsibleAlignmentProps {
     | Array<ReactElement<ColumnProps> | null>
     | ReactElement<ColumnProps>
     | null;
+  data?: DataAttributeMap;
 }
 
 export const Columns = ({
@@ -49,6 +53,7 @@ export const Columns = ({
   space = 'none',
   align,
   alignY,
+  data,
 }: ColumnsProps) => {
   const normalizedSpace = normalizeResponsiveValue(space);
   const {
@@ -77,7 +82,11 @@ export const Columns = ({
   });
 
   return (
-    <Box {...collapsibleAlignmentProps} className={negativeMarginLeft}>
+    <Box
+      {...collapsibleAlignmentProps}
+      className={negativeMarginLeft}
+      {...(data ? buildDataAttributes(data) : undefined)}
+    >
       <ColumnsContext.Provider
         value={{
           collapseMobile,