+++ lib/components/Hidden/Hidden.tsx
@@ -6,6 +6,9 @@ import {
   resolveResponsiveRangeProps,
   ResponsiveRangeProps,
 } from '../../utils/responsiveRangeProps';
+import buildDataAttributes, {
+  DataAttributeMap,
+} from '../private/buildDataAttributes';
 import * as styles from './Hidden.css';
 
 export interface HiddenProps extends ResponsiveRangeProps {
@@ -14,6 +17,7 @@ export interface HiddenProps extends ResponsiveRangeProps {
   screen?: boolean;
   print?: boolean;
   inline?: boolean;
+  data?: DataAttributeMap;
 }
 
 export const Hidden = ({
@@ -24,6 +28,7 @@ export const Hidden = ({
   screen,
   print,
   inline: inlineProp,
+  data,
 }: HiddenProps) => {
   if (process.env.NODE_ENV === 'development' && screen) {
     // eslint-disable-next-line no-console
@@ -59,6 +64,7 @@ export const Hidden = ({
       }
       className={hiddenOnPrint ? styles.hiddenOnPrint : undefined}
       component={component || (inline ? 'span' : 'div')}
+      {...(data ? buildDataAttributes(data) : undefined)}
     >
       {children}
     </Box>