+++ lib/components/MonthPicker/MonthPicker.tsx
@@ -52,7 +52,6 @@ export interface MonthPickerProps
     | 'value'
     | 'labelId'
     | 'aria-describedby'
-    | 'data'
     | 'name'
     | 'autoComplete'
     | 'secondaryMessage'
@@ -199,7 +198,6 @@ const MonthPicker = ({
       icon={undefined}
       prefix={undefined}
       labelId={undefined}
-      data={undefined}
       name={undefined}
       autoComplete={undefined}
       secondaryMessage={null}