+++ lib/components/RadioGroup/RadioItem.tsx
@@ -21,6 +21,7 @@ export interface RadioItemProps
     | 'id'
     | 'disabled'
     | 'tone'
+    | 'size'
   > {
   value: NonNullable<InlineFieldProps['value']>;
 }
@@ -58,6 +59,7 @@ const NamedRadioItem = forwardRef<HTMLInputElement, RadioItemProps>(
             ? radioGroupContext.tone
             : undefined
         }
+        size={radioGroupContext.size}
         disabled={radioGroupContext.disabled}
         aria-describedby={radioGroupContext['aria-describedby']}
         tabIndex={tababble ? 0 : -1}