+++ lib/components/Tabs/TabPanel.tsx
@@ -45,7 +45,7 @@ export const TabPanel = ({ children, data, item }: TabPanelProps) => {
       position="relative"
       outline="none"
       className={styles.tabPanel}
-      {...buildDataAttributes(data)}
+      {...(data ? buildDataAttributes(data) : undefined)}
     >
       {isSelected || renderInactive ? children : undefined}
       <Overlay