+++ lib/components/Toggle/Toggle.tsx
@@ -5,7 +5,11 @@ import { Text } from '../Text/Text';
 import { IconTick } from '../icons';
 import { useVirtualTouchable } from '../private/touchable/useVirtualTouchable';
 import { useBackgroundLightness } from '../Box/BackgroundContext';
+import buildDataAttributes, {
+  DataAttributeMap,
+} from '../private/buildDataAttributes';
 import * as styles from './Toggle.css';
+import type { Size } from './Toggle.css';
 
 type HTMLInputProps = AllHTMLAttributes<HTMLInputElement>;
 type ChangeHandler = (value: boolean) => void;
@@ -15,6 +19,8 @@ export interface ToggleProps {
   on: boolean;
   onChange: ChangeHandler;
   align?: 'left' | 'right' | 'justify';
+  size?: Size;
+  data?: DataAttributeMap;
 }
 
 const handleChange = (onChange: ChangeHandler) => (
@@ -31,6 +37,8 @@ export const Toggle = ({
   onChange,
   label,
   align = 'left',
+  size = 'standard',
+  data,
 }: ToggleProps) => {
   const showBorder = useBackgroundLightness() === 'light';
 
@@ -41,6 +49,7 @@ export const Toggle = ({
       display="flex"
       flexDirection={align === 'left' ? undefined : 'rowReverse'}
       className={styles.root}
+      {...(data ? buildDataAttributes(data) : undefined)}
     >
       <Box
         component="input"
@@ -52,20 +61,28 @@ export const Toggle = ({
         zIndex={1}
         cursor="pointer"
         opacity={0}
-        className={[styles.realField, styles.fieldSize]}
+        className={[
+          styles.realField,
+          styles.realFieldPosition[size],
+          styles.fieldSize[size],
+        ]}
       />
       <Box
         position="relative"
         display="flex"
         alignItems="center"
         flexShrink={0}
-        className={[styles.slideContainer, styles.fieldSize]}
+        className={[
+          styles.slideContainerBase,
+          styles.slideContainerSize[size],
+          styles.fieldSize[size],
+        ]}
       >
         <Box
           position="absolute"
           width="full"
           overflow="hidden"
-          className={styles.slideTrack}
+          className={[styles.slideTrack[size], styles.slideTrackBackground]}
         >
           <Box
             position="absolute"
@@ -85,7 +102,7 @@ export const Toggle = ({
           alignItems="center"
           justifyContent="center"
           borderRadius="full"
-          className={styles.slider}
+          className={styles.slider[size]}
         >
           <FieldOverlay className={styles.icon}>
             <IconTick tone="formAccent" size="fill" />
@@ -112,9 +129,9 @@ export const Toggle = ({
         flexGrow={align === 'justify' ? 1 : undefined}
         userSelect="none"
         cursor="pointer"
-        className={[styles.label, useVirtualTouchable()]}
+        className={[styles.label[size], useVirtualTouchable()]}
       >
-        <Text baseline={false} weight={on ? 'strong' : undefined}>
+        <Text baseline={false} weight={on ? 'strong' : undefined} size={size}>
           {label}
         </Text>
       </Box>