+++ lib/components/private/InlineField/InlineField.css.ts
@@ -1,8 +1,15 @@
-import { style } from '@vanilla-extract/css';
+import { style, styleVariants } from '@vanilla-extract/css';
 import { calc } from '@vanilla-extract/css-utils';
 import { themeVars } from '../../../themes/themeVars.css';
 import { hitArea } from '../touchable/hitArea';
 import { debugTouchable } from '../touchable/debugTouchable';
+
+const sizes = {
+  standard: 'standard',
+  small: 'small',
+} as const;
+
+export type Size = keyof typeof sizes;
 
 // Reset the z-index at the parent level to scope
 // overrides internally.
@@ -12,48 +19,58 @@ export const root = style({
   },
 });
 
-const centerOffsetCalc = calc(hitArea)
-  .subtract(themeVars.inlineFieldSize)
-  .divide(2)
-  .negate()
-  .toString();
-
 export const realField = style({
   width: hitArea,
   height: hitArea,
-  top: centerOffsetCalc,
-  left: centerOffsetCalc,
   selectors: {
     ...debugTouchable(),
   },
 });
 
-export const fakeField = style({
-  height: themeVars.inlineFieldSize,
-  width: themeVars.inlineFieldSize,
-});
-
-const badgeOffsetCalc = calc(themeVars.inlineFieldSize)
-  .subtract(themeVars.typography.text.xsmall.mobile.leading)
-  .divide(2)
-  .toString();
-
-export const badgeOffset = style({
-  paddingTop: badgeOffsetCalc,
-  paddingBottom: badgeOffsetCalc,
-});
-
-export const label = style({
-  paddingTop: calc(themeVars.inlineFieldSize)
-    .subtract(themeVars.typography.text.standard.mobile.capHeight)
+export const realFieldPosition = styleVariants(sizes, (size: Size) => {
+  const offset = calc(hitArea)
+    .subtract(themeVars.inlineFieldSize[size])
     .divide(2)
-    .toString(),
+    .negate()
+    .toString();
+
+  return {
+    top: offset,
+    left: offset,
+  };
+});
+
+export const fakeFieldBase = style({});
+export const fakeFieldSize = styleVariants(sizes, (size) => ({
+  height: themeVars.inlineFieldSize[size],
+  width: themeVars.inlineFieldSize[size],
+}));
+
+export const badgeOffset = styleVariants(sizes, (size: Size) => {
+  const offset = calc(themeVars.inlineFieldSize[size])
+    .subtract(themeVars.typography.text.xsmall.mobile.leading)
+    .divide(2)
+    .toString();
+
+  return {
+    paddingTop: offset,
+    paddingBottom: offset,
+  };
+});
+
+export const labelBase = style({
   selectors: {
     [`${realField}:not(:disabled) ~ * &`]: {
       cursor: 'pointer',
     },
   },
 });
+export const labelOffset = styleVariants(sizes, (size: Size) => ({
+  paddingTop: calc(themeVars.inlineFieldSize[size])
+    .subtract(themeVars.typography.text[size].mobile.capHeight)
+    .divide(2)
+    .toString(),
+}));
 
 export const isMixed = style({});
 
@@ -67,7 +84,7 @@ export const children = style({
 
 export const selected = style({
   selectors: {
-    [`${realField}:checked + ${fakeField} > &, ${realField}${isMixed} + ${fakeField} > &`]: {
+    [`${realField}:checked + ${fakeFieldBase} > &, ${realField}${isMixed} + ${fakeFieldBase} > &`]: {
       opacity: 1,
     },
   },
@@ -75,7 +92,7 @@ export const selected = style({
 
 export const focusOverlay = style({
   selectors: {
-    [`${realField}:focus + ${fakeField} > &`]: {
+    [`${realField}:focus + ${fakeFieldBase} > &`]: {
       opacity: 1,
     },
   },
@@ -83,7 +100,7 @@ export const focusOverlay = style({
 
 export const hoverOverlay = style({
   selectors: {
-    [`${realField}:hover:not(:checked):not(${isMixed}):not(:disabled) + ${fakeField} > &`]: {
+    [`${realField}:hover:not(:checked):not(${isMixed}):not(:disabled) + ${fakeFieldBase} > &`]: {
       opacity: 1,
     },
   },
@@ -100,7 +117,7 @@ export const indicator = style({
 const checkboxScale = style({
   transform: 'scale(0.85)',
   selectors: {
-    [`${realField}:active + ${fakeField} > * > &`]: {
+    [`${realField}:active + ${fakeFieldBase} > * > &`]: {
       transform: 'scale(0.75)',
     },
   },
@@ -111,7 +128,7 @@ export const checkboxIndicator = [indicator, checkboxScale];
 const radioScale = style({
   transform: 'scale(0.6)',
   selectors: {
-    [`${realField}:active + ${fakeField} > * > &`]: {
+    [`${realField}:active + ${fakeFieldBase} > * > &`]: {
       transform: 'scale(0.5)',
     },
   },